move page objects to match views
